(many): Various changes
These changes have been made as part of the move towards adding more
planning strategies.

-	`Schedule` is now a trait. The original `Schedule` struct is now
	`PomodoroSchedule`, and it itself implements the new `Schedule`
	trait.

-	`schedule` in the config file is now optional and has been
	renamed to `pomodoro_schedule`.

-	`ScheduleEntry` is now specific to the pomodoro schedule

-	Any iterators producing `ScheduleEntry`s are also specific to
	the pomodoro implementation

Changes to `TimeSpan`:

-	`beginning()` is now `start()`, because it
	just makes more sense (i guess). It's more concise and less
	prone to haphazard spelling errors (because I can never remember
	how to spell "beggining"

-	`set_start` allows the start of a `TimeSpan` to be changed

Next steps (TODO):

-	Break up the pomodoro module into smaller files because holy
	moly

-	Allow unlimited arbitrary creation of schedules of all types. A
	new enum `ScheduleType` will allow a `HashMap<String,
	ScheduleType>` to be deserialized and serialized as part of the
	user's data file. User-specified schedule names (the `String`
	key) set each schedule apart.

-	Implement the Ivy Lee planning method, which literally just
	takes the next six (sub)tasks due and plans them (with no specific
	time slots) for the next day
