+++ TODO.md
@@ -2,5 +2,3 @@ TODO
 ============================================================
 
 -   Allow partial WorkSessions
-
-
