+++ src/methods/ivy_lee.rs
@@ -1,0 +1,1 @@
+pub struct IvyLeeSchedule;
